Move utilities to separate module
